We can simplify the for loop while still being readable
That for loop does three things, it iterates over the numbers making a new number, it adds up the new numbers and it works out the positive difference.
The positive difference is also just the absolute of a-b.
Iterating over all the numbers creating those totals is an easy map, and then we add them together