Ok day 1 part 1, that's just nasty.
The generator and even the filter on all isn't too bad.  I used a let diff = line because it was easier to read the actual test that way.

But the use of -99 as a flag that I'd found a change of direction feels wrong.

But it works and it's reasonably simple