Refactor to tidy up
Jeff in the cross gov slack reminded me that Enums can have member variables.  That means that instead of passing tuples around with two unused numbers, I can embed the parameters to MUL in the enum, removing all the 0,0's.  It also means we can use pattern matching and decomposition to better name the variables