I've fought the compiler and the compiler won.
An hour of faffing with compiler errors.  Turning a string into a vector of characters is a pain, converting usize to i32 is a pain, and copying vectors to vectors is a pain.

but I can now read the input and turn it into a generator result that I think I can do stuff too