Sigh, Day 4, Part 4 - Harder if you forget something critical
I did this fiarly quickly, but my tests kept giving the wrong answers.  I changed approach, and that also gave me the wrong answer.

Turns out if that if you are looking for the center of the MAS star, you need to look for A not X as the central character.

Otherwise happy with this I think