Stopping here for debug purposes
Interesting edge case with one of the test cases.  The count is supposed to be the number of summits that a trailhead can reach, but I'm getting a wrong number.  Some debugging, and in the test case, there are multiple ways to get from 0x6 to 3,0, some 9 in fact.

Part 1 only wants the reachable index, but I strongly suspect that Part 2 might want us to count unique paths, so I want to keep this "bug", but for now, I think we probably actually want to count summit,trailhead pairs instead.